pyproject.toml
[tool.setuptools]
packages = ["shared", "models"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pybinbot"
version = "0.0.1"
description = "Utility functions for the binbot project."
authors = [
    { name="Carlos Wu", email="carkodw@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiokafka>=0.12.0",
    "apscheduler>=3.6.3",
    "confluent-kafka>=2.11.1",
    "kafka-python>=2.0.2",
    "kucoin-universal-sdk>=1.3.0",
    "numpy==2.2.0",
    "pandas>=2.2.3",
    "passlib",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.6.3",
    "python-dotenv",
    "python-jose",
    "requests>=2.28.1",
    "requests-cache>=1.2.0",
    "requests-html>=0.10.0",
    "scipy==1.14.1",
    "websocket-client>=1.5.0",
]
